+++ src/Pages/Works/index.tsx
@@ -8,7 +8,7 @@ import { withStyles, WithStyles, createStyles } from '@material-ui/core/styles'
 import React, { Component } from 'react'
 import { Helmet } from 'react-helmet'
 
-import PageTitleWrapper from '../../Shared/Styles/PageTitleWrapper'
+import PageTitleWrapper from '../../Components/PageTitleWrapper'
 import contents from './WorksContents.json'
 
 const styleSettings = () =>
